% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_authorize_account.R
\name{b2_authorize_account}
\alias{b2_authorize_account}
\title{Authorize Backblaze B2}
\usage{
b2_authorize_account(account_id = Sys.getenv("B2_ACCOUNT_ID"),
  application_key = Sys.getenv("B2_APPLICATION_KEY"))
}
\arguments{
\item{account_id}{B2 account ID}

\item{application_key}{B2 application key for the account}
}
\value{
Invisibly, a list with the values from the JSON response
}
\description{
Log in to B2 API
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_authorize_account.html}
}
