% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_delete_file_version.R
\name{b2_delete_file_by_name}
\alias{b2_delete_file_by_name}
\title{Delete file version}
\usage{
b2_delete_file_by_name(bucket_name, file_name)
}
\arguments{
\item{bucket_name}{The name of the bucket containing the file}

\item{file_name}{The file name}
}
\value{
A list of the JSON response
}
\description{
Delete latest version of a file stored in B2 by name
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_delete_file_version.html}

To delete a file using the file name and file id, use
\code{\link{b2_delete_file_by_name}}.
}
