% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_delete_file_version.R
\name{b2_delete_file_version}
\alias{b2_delete_file_version}
\title{Delete file version}
\usage{
b2_delete_file_version(file_name, file_id)
}
\arguments{
\item{file_name}{The file name}

\item{file_id}{The file id}
}
\value{
A list of the JSON response
}
\description{
Delete latest version of a file stored in B2
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_delete_file_version.html}

Note that the file id is required in addition to the file name. To
delete a file using the bucket name and file name, use
\code{\link{b2_delete_file_by_name}}.
}
