% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_download_files.R
\name{b2_download_file_by_id}
\alias{b2_download_file_by_id}
\title{Download a file}
\usage{
b2_download_file_by_id(file_id, output_file, overwrite = FALSE, ...)
}
\arguments{
\item{file_id}{The file id}

\item{output_file}{(Optional) Path of the file on disk to which to save
the downloaded content}

\item{overwrite}{Whether to overwrite the existing output file}

\item{...}{Optional parameters (see details)}
}
\value{
The response from \code{httr::GET}
}
\description{
Download a file in B2 by the file id
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_download_file_by_id.html}

The file id can be obtained from a call to \code{\link{b2_list_file_names}}.

To download a file using the bucket name and file use, call
\code{\link{b2_download_file_by_name}}

\subsection{Optional parameters}{
An optional arguments to the API call can be passed as named arguments

\code{Range} - A byte-range request, as described in
\url{https://www.backblaze.com/b2/docs/b2_download_file_by_id.html}

}
}
