% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_download_files.R
\name{b2_download_file_by_name}
\alias{b2_download_file_by_name}
\title{Download a file}
\usage{
b2_download_file_by_name(bucket_name, file_name, output_file,
  overwrite = FALSE, ...)
}
\arguments{
\item{bucket_name}{The name of the bucket containing the file}

\item{file_name}{The name of the file in the bucket}

\item{output_file}{(Optional) Path of the file on disk to which to save
the downloaded content}

\item{overwrite}{Whether to overwrite the existing output file}

\item{...}{Optional parameters (see details)}
}
\value{
The response from \code{httr::GET}
}
\description{
Download a file in B2 by the file name
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_download_file_by_name.html}

To download a file using the file id, call
\code{\link{b2_download_file_by_id}}

\subsection{Optional parameters}{
An optional arguments to the API call can be passed as named arguments

\code{Range} - A byte-range request, as described in
\url{https://www.backblaze.com/b2/docs/b2_download_file_by_id.html}

}
}
