% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_list_file_names.R
\name{b2_files_by_bucket_name}
\alias{b2_files_by_bucket_name}
\alias{b2_files_by_bucket_name_df}
\title{List file names}
\usage{
b2_files_by_bucket_name(bucket_name, account_id = Sys.getenv("B2_ACCOUNT_ID"),
  ...)

b2_files_by_bucket_name_df(bucket_name,
  account_id = Sys.getenv("B2_ACCOUNT_ID"), ...)
}
\arguments{
\item{bucket_name}{Name of the bucket to get list of files}

\item{account_id}{B2 account ID}

\item{...}{Optional parameters (see details)}
}
\value{
\code{b2_files_by_bucket_name} returns a list of the JSON response.

\code{b2_files_by_bucket_name_df} returns a tibble of the file information.
}
\description{
List of files in a B2 bucket, starting at a given file name
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_list_file_names.html}

In this function, you provide the name of the bucket, as opposed to
\code{\link{b2_list_file_names}}, where you provide the id of the bucket.

\subsection{Optional parameters}{
Optional arguments to the API call can be passed as named arguments

\code{startFileName} - The first file name to return

\code{maxFileCount} - Maximum number of files to return (defaults to 100, maximum 10000)

\code{prefix} - Return files starting with prefix

\code{delimiter} - Used to limit return to top folder or single subfolder
}
}
