% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_get_file_info.R
\name{b2_get_file_info}
\alias{b2_get_file_info}
\title{Get file info}
\usage{
b2_get_file_info(file_id)
}
\arguments{
\item{file_id}{The file id}
}
\value{
A list of the JSON response
}
\description{
Get information about a file stored in B2
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_get_file_info.html}

Note that the file id, not the file name, is required. To get file
information using the bucket name and file name, use
\code{\link{b2_get_file_info_by_name}}.
}
