% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_get_file_info.R
\name{b2_get_file_info_by_name}
\alias{b2_get_file_info_by_name}
\title{Get file info}
\usage{
b2_get_file_info_by_name(bucket_name, file_name)
}
\arguments{
\item{bucket_name}{The name of the bucket containing the file}

\item{file_name}{The name of the file}
}
\value{
A list of the JSON response
}
\description{
Get information about a file stored in B2
}
\details{
To get file information using the file id, use \code{\link{b2_get_file_info}}.
}
