% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_upload_file.R
\name{b2_get_upload_url}
\alias{b2_get_upload_url}
\title{Get upload URL}
\usage{
b2_get_upload_url(bucket_id)
}
\arguments{
\item{bucket_name}{The name of the bucket containing the file}

\item{file_name}{The name of the file}
}
\value{
A list of the JSON response
}
\description{
Get a URL for uploading a file to B2
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_get_upload_url.html}
}
