% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_list_buckets.R
\name{b2_list_buckets}
\alias{b2_list_buckets}
\alias{b2_list_buckets_df}
\title{List B2 buckets}
\usage{
b2_list_buckets(account_id = Sys.getenv("B2_ACCOUNT_ID"))

b2_list_buckets_df(account_id = Sys.getenv("B2_ACCOUNT_ID"))
}
\arguments{
\item{account_id}{B2 account ID}
}
\value{
\code{b2_list_buckets} returns a list from the JSON reponse.

\code{b2_list_buckets_df} returns a tibble with information about the buckets
}
\description{
List buckets associated with an account
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_list_buckets.html}
}
