% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_list_file_names.R
\name{b2_list_file_names}
\alias{b2_list_file_names}
\alias{b2_list_file_names_df}
\title{List file names}
\usage{
b2_list_file_names(bucket_id, ...)

b2_list_file_names_df(bucket_id, ...)
}
\arguments{
\item{bucket_id}{Id of the bucket to get list of files. Note that this is not the name
of the bucket.}

\item{...}{Optional parameters (see details)}
}
\value{
\code{b2_list_file_names} returns a list of the JSON response.

\code{b2_list_file_names_df} returns a tibble of the file information.
}
\description{
List of files in a B2 bucket, starting at a given file name
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_list_file_names.html}

Note that you provide the id of the bucket to this function, not the name
of the bucket. The id of the bucket is obtained from a call to
\code{\link{b2_list_buckets}}.

To get the a list of files by bucket name, call \code{\link{b2_files_by_bucket_name}}

\subsection{Optional parameters}{
Optional arguments to the API call can be passed as named arguments

\code{startFileName} - The first file name to return

\code{maxFileCount} - Maximum number of files to return (defaults to 100, maximum 10000)

\code{prefix} - Return files starting with prefix

\code{delimiter} - Used to limit return to top folder or single subfolder
}
}
