% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_upload_file.R
\name{b2_upload_file}
\alias{b2_upload_file}
\title{Upload a file}
\usage{
b2_upload_file(data, upload_url, upload_token, upload_filename,
  mime_type = mime::guess_type(upload_filename))
}
\arguments{
\item{data}{Either a raw vector of the data or the path to the file}

\item{upload_url}{The upload_url}

\item{upload_token}{The upload token}

\item{upload_filename}{Name of the file to save in B2}

\item{mime_type}{MIME type}
}
\value{
A list of the JSON response
}
\description{
Upload a file to B2
}
\details{
See \url{https://www.backblaze.com/b2/docs/b2_upload_file.html}

This requires getting an upload URL and token by calling
\code{\link{b2_get_upload_url}} first

To upload a file directly, use \code{\link{b2_upload_file_direct}}
}
