% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2_upload_file.R
\name{b2_upload_file_direct}
\alias{b2_upload_file_direct}
\title{Upload a file}
\usage{
b2_upload_file_direct(data, bucket_id, upload_filename,
  mime_type = mime::guess_type(upload_filename))
}
\arguments{
\item{data}{Either a raw vector of the data or the path to the file}

\item{bucket_id}{The id of the bucket to which to upload the file}

\item{upload_filename}{Name of the file to save in B2}

\item{mime_type}{MIME type}
}
\value{
A list of the JSON response
}
\description{
Upload a file to B2
}
\details{
This performs both the step of getting an upload URL (\code{\link{b2_get_upload_url}})
and uploading the file (\code{\link{b2_upload_file}})
}
